// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-workspace-delta-response-api-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingWorkspaceDeltaResponse } from '../models/linking-workspace-delta-response';
export interface LinkingWorkspaceDeltaResponseApiResponse {
  data: LinkingWorkspaceDeltaResponse | null;
  errors: Array<string> | null;
  isSuccess: boolean;
  message: string | null;
}